Demodulize annotation resources
